# PERF: using murmur hash for float64 khash-tables