@realead 2020-10-14
Now the comparison of different probing strategies:

    the base: double hashing (the state of this PR)
    linear probing (realead@1dc7af7)
    quadratic probing (i.e. khash 0.2.8) (realead@59600e5)
    combined linear probing + double hashing (realead@5e44098)

The different probing strategies are interesting for cases where comparisons are cheap (thus cache misses play a role), e.g. float64/int64/uint64. For heavier types (PyObject, strings), it plays only a role when there are problems with robustness in such a a way that look-up costs more than `O(n)`.

Linear probing:

Linear probing will be really bad for types with bad first (and only) hash-function at least for some series. This is what we see (see the whole comparison at the end of comment):
 
  * hash for `PyObject` is really bad, some test cases become more than 100 times slower
  * hash for int/uint is better, but there is still a "bad" series for which it becomes factor 20 slower
  * the used murmur2 hash for float64 is quite strong, no problematic series in the test suite.

_Quadratic probing_

Supposed to be better than linear probing, as more robust. This is what we also can observe (all results at the end):

 * PyObject is only 20 times slower
 * problematic series for int64/uint64 are only 2 times slower
 * otherwise gains are similar to linear

_Combined probing:_

Theoretically best of both worlds: robust as double hashing but as few cache misses as quadratic probing. It looks as expected: there are some cases with about 10% slow-down but many  more cases with speed-ups, comparable with speed-ups from quadratic probing. Also for weak hashes nothing bad happens. Here are the all comparison for combined probing:

```
      before           after         ratio
     [c852b2ee]       [5e44098d]
+     23.0±0.07ms       27.9±0.9ms     1.21  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 1000, 2)
+      1.93±0.2μs       2.29±0.2μs     1.18  index_cached_properties.IndexCache.time_inferred_type('DatetimeIndex')
+      24.7±0.3ms       28.5±0.7ms     1.15  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 1000, 2)
+      23.7±0.1ms       27.1±0.5ms     1.15  hash_functions.IsinWithArange.time_isin(<class 'numpy.float64'>, 8000, 2)
+      1.83±0.1μs       2.08±0.2μs     1.14  index_cached_properties.IndexCache.time_values('DatetimeIndex')
+      23.5±0.3ms       26.6±0.7ms     1.13  hash_functions.IsinWithArange.time_isin(<class 'numpy.float64'>, 8000, -2)
+      23.5±0.2ms       26.4±0.5ms     1.12  hash_functions.IsinWithArange.time_isin(<class 'numpy.float64'>, 1000, -2)
+      23.4±0.2ms       26.2±0.5ms     1.12  hash_functions.IsinWithArange.time_isin(<class 'numpy.float64'>, 2000, 2)
+      12.6±0.1ms       14.1±0.5ms     1.12  indexing.IntervalIndexing.time_loc_list
+     4.98±0.02ms      5.57±0.05ms     1.12  timeseries.ResampleSeries.time_resample('period', '1D', 'mean')
+      23.2±0.2ms       25.9±0.8ms     1.11  hash_functions.IsinWithArange.time_isin(<class 'numpy.float64'>, 2000, -2)
+        664±50ns         739±70ns     1.11  index_cached_properties.IndexCache.time_inferred_type('RangeIndex')
+     1.18±0.02ms      1.32±0.01ms     1.11  series_methods.IsInForObjects.time_isin_short_series_long_values
+      1.78±0.1μs       1.98±0.2μs     1.11  index_cached_properties.IndexCache.time_values('PeriodIndex')
+      3.66±0.2μs       4.05±0.4μs     1.11  index_cached_properties.IndexCache.time_inferred_type('Float64Index')
+     4.28±0.04ms      4.72±0.08ms     1.10  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'numpy.float64'>, 80000)
-     3.34±0.05μs      3.04±0.02μs     0.91  tslibs.tslib.TimeIntsToPydatetime.time_ints_to_pydatetime('time', 0, tzlocal())
-        16.9±1μs       15.4±0.2μs     0.91  dtypes.Dtypes.time_pandas_dtype('timedelta64')
-      1.05±0.2ms         949±60μs     0.91  ctors.SeriesConstructors.time_series_constructor(<function list_of_str at 0x7f8dd9d979d0>, False, 'int')
-      5.01±0.1ms      4.54±0.03ms     0.91  index_object.Indexing.time_get_loc_non_unique('Float')
-      3.08±0.3ms      2.79±0.07ms     0.91  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 1, 'linear')
-        705±10ms          638±5ms     0.90  join_merge.I8Merge.time_i8merge('outer')
-         727±4ms          656±4ms     0.90  join_merge.I8Merge.time_i8merge('left')
-     3.19±0.05μs      2.88±0.02μs     0.90  tslibs.tslib.TimeIntsToPydatetime.time_ints_to_pydatetime('time', 0, datetime.timezone.utc)
-     2.79±0.03ms      2.52±0.02ms     0.90  indexing.NumericSeriesIndexing.time_loc_list_like(<class 'pandas.core.indexes.numeric.Float64Index'>, 'nonunique_monotonic_inc')
-      34.8±0.5ms       31.4±0.7ms     0.90  hash_functions.IsinAlmostFullWithRandomInt.time_isin(<class 'numpy.uint64'>, 20)
-         393±4ms          354±5ms     0.90  hash_functions.NumericSeriesIndexingShuffled.time_loc_slice(<class 'pandas.core.indexes.numeric.Int64Index'>, 5000000)
-     3.34±0.07μs      3.00±0.06μs     0.90  tslibs.tslib.TimeIntsToPydatetime.time_ints_to_pydatetime('time', 0, None)
-         311±3μs          279±3μs     0.90  hash_functions.IsinWithArangeSorted.time_isin(<class 'numpy.int64'>, 8000)
-        3.51±1ms       3.14±0.1ms     0.90  ctors.SeriesConstructors.time_series_constructor(<class 'list'>, False, 'int')
-      1.44±0.5ms      1.29±0.08ms     0.90  frame_ctor.FromRecords.time_frame_from_records_generator(1000)
-         113±1ms        101±0.8ms     0.90  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'numpy.float64'>, 900000)
-        72.1±2ms       64.4±0.9ms     0.89  hash_functions.IsinAlmostFullWithRandomInt.time_isin(<class 'numpy.float64'>, 20)
-        13.3±1μs       11.9±0.6μs     0.89  index_cached_properties.IndexCache.time_is_all_dates('TimedeltaIndex')
-     3.70±0.08μs      3.29±0.05μs     0.89  tslibs.tslib.TimeIntsToPydatetime.time_ints_to_pydatetime('date', 1, datetime.timezone.utc)
-      3.32±0.8ms       2.95±0.1ms     0.89  ctors.SeriesConstructors.time_series_constructor(<function arr_dict at 0x7f8dd9d97af0>, False, 'float')
-        88.9±1ms         79.0±1ms     0.89  multiindex_object.Duplicated.time_duplicated
-         394±2ms          349±4ms     0.89  hash_functions.NumericSeriesIndexingShuffled.time_loc_slice(<class 'pandas.core.indexes.numeric.UInt64Index'>, 5000000)
-      12.5±0.6ms       11.1±0.6ms     0.89  hash_functions.UniqueAndFactorizeArange.time_unique(10)
-         753±2ms          667±6ms     0.88  join_merge.I8Merge.time_i8merge('right')
-      4.20±0.1μs      3.72±0.06μs     0.88  tslibs.tslib.TimeIntsToPydatetime.time_ints_to_pydatetime('datetime', 1, None)
-      3.14±0.3ms      2.77±0.03ms     0.88  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0, 'higher')
-        42.4±2ms       37.4±0.4ms     0.88  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0.5, 'midpoint')
-         138±5ms          122±3ms     0.88  gil.ParallelGroupbyMethods.time_loop(8, 'last')
-         196±2ms          173±2ms     0.88  sparse.SparseSeriesToFrame.time_series_to_frame
-        35.1±4ms       30.9±0.8ms     0.88  gil.ParallelGroupbyMethods.time_loop(2, 'sum')
-      32.0±0.8ms       28.1±0.4ms     0.88  hash_functions.IsinAlmostFullWithRandomInt.time_isin(<class 'numpy.int64'>, 20)
-      12.4±0.6ms       10.9±0.4ms     0.88  hash_functions.UniqueAndFactorizeArange.time_unique(15)
-        137±10ms          120±1ms     0.88  gil.ParallelGroupbyMethods.time_loop(8, 'prod')
-        12.7±2ms       11.0±0.5ms     0.87  hash_functions.UniqueAndFactorizeArange.time_unique(11)
-        958±10μs         834±30μs     0.87  timeseries.DatetimeIndex.time_unique('repeated')
-      1.01±0.3ms         876±40μs     0.87  ctors.SeriesConstructors.time_series_constructor(<function list_of_str at 0x7f8dd9d979d0>, False, 'float')
-        29.1±2μs         25.3±2μs     0.87  index_cached_properties.IndexCache.time_engine('CategoricalIndex')
-      12.5±0.4ms       10.9±0.4ms     0.87  hash_functions.UniqueAndFactorizeArange.time_unique(4)
-        40.1±3ms       34.8±0.3ms     0.87  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0.5, 'higher')
-      1.04±0.2ms         903±50μs     0.87  ctors.SeriesConstructors.time_series_constructor(<function list_of_str at 0x7f8dd9d979d0>, True, 'float')
-         141±6ms          122±1ms     0.87  gil.ParallelGroupbyMethods.time_loop(8, 'min')
-      1.44±0.2ms      1.25±0.08ms     0.87  ctors.SeriesConstructors.time_series_constructor(<class 'list'>, False, 'float')
-       94.5±20μs        81.7±10μs     0.86  ctors.SeriesConstructors.time_series_constructor(<function no_change at 0x7f8dd9d97940>, False, 'float')
-        72.5±1ms         62.6±1ms     0.86  hash_functions.IsinWithRandomFloat.time_isin(<class 'numpy.float64'>, 750000)
-        3.64±1ms       3.13±0.1ms     0.86  ctors.SeriesConstructors.time_series_constructor(<function arr_dict at 0x7f8dd9d97af0>, True, 'float')
-        580±10ms          499±6ms     0.86  hash_functions.NumericSeriesIndexing.time_loc_slice(<class 'pandas.core.indexes.numeric.Float64Index'>, 5000000)
-        40.2±2ms       34.5±0.3ms     0.86  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0.5, 'nearest')
-      83.8±0.6ms       71.9±0.3ms     0.86  series_methods.IsInFloat64.time_isin_many_different
-     3.41±0.03ms      2.92±0.03ms     0.86  indexing.NumericSeriesIndexing.time_getitem_list_like(<class 'pandas.core.indexes.numeric.Float64Index'>, 'nonunique_monotonic_inc')
-      12.6±0.5ms       10.8±0.4ms     0.85  hash_functions.UniqueAndFactorizeArange.time_unique(8)
-         106±1ms         89.9±2ms     0.85  join_merge.Align.time_series_align_int64_index
-        580±10ms          492±4ms     0.85  hash_functions.NumericSeriesIndexingShuffled.time_loc_slice(<class 'pandas.core.indexes.numeric.Float64Index'>, 5000000)
-        41.2±3ms       34.7±0.2ms     0.84  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0.5, 'lower')
-      1.53±0.5ms      1.28±0.09ms     0.84  ctors.SeriesConstructors.time_series_constructor(<class 'list'>, True, 'float')
-        76.7±4ms         63.5±2ms     0.83  gil.ParallelGroupbyMethods.time_loop(4, 'max')
-      2.69±0.5ms      2.23±0.03ms     0.83  frame_methods.Lookup.time_frame_fancy_lookup
-         123±1ms          102±2ms     0.83  hash_functions.IsinWithRandomFloat.time_isin(<class 'numpy.float64'>, 900000)
-      1.96±0.1ms      1.62±0.03ms     0.83  dtypes.SelectDtypes.time_select_dtype_float_exclude('UInt8')
-      73.8±0.9ms         61.0±1ms     0.83  gil.ParallelGroupbyMethods.time_loop(4, 'sum')
-         134±2ms          111±1ms     0.83  hash_functions.IsinWithArangeSorted.time_isin(<class 'numpy.float64'>, 1000000)
-        74.7±4ms         61.4±1ms     0.82  gil.ParallelGroupbyMethods.time_loop(4, 'last')
-      92.0±0.8ms         75.2±1ms     0.82  join_merge.Align.time_series_align_left_monotonic
-        120±40μs         97.3±8μs     0.81  ctors.SeriesConstructors.time_series_constructor(<function no_change at 0x7f8dd9d97940>, True, 'float')
-      8.13±0.1ms      6.59±0.09ms     0.81  index_object.SetOperations.time_operation('int', 'symmetric_difference')
-      13.4±0.6ms       10.9±0.4ms     0.81  hash_functions.UniqueAndFactorizeArange.time_unique(12)
-       659±200ms          533±9ms     0.81  categoricals.Indexing.time_reindex_missing
-        7.14±1ms      5.71±0.07ms     0.80  algorithms.Factorize.time_factorize(False, True, 'boolean')
-      2.03±0.4ms      1.62±0.03ms     0.79  arithmetic.NumericInferOps.time_divide(<class 'numpy.uint8'>)
-       77.0±20ms       61.2±0.8ms     0.79  gil.ParallelGroupbyMethods.time_loop(4, 'mean')
-        7.72±1μs       6.11±0.5μs     0.79  index_cached_properties.IndexCache.time_is_all_dates('DatetimeIndex')
-      5.39±0.8ms       4.25±0.2ms     0.79  algorithms.Factorize.time_factorize(True, False, 'datetime64[ns]')
-        151±20ms          119±2ms     0.79  frame_methods.Duplicated.time_frame_duplicated
-      3.55±0.3ms      2.79±0.06ms     0.79  rolling.Quantile.time_quantile('DataFrame', 10, 'int', 0, 'midpoint')
-        37.6±4ms       29.5±0.4ms     0.79  frame_ctor.FromLists.time_frame_from_lists
-      9.55±0.1ms       7.44±0.1ms     0.78  index_object.SetOperations.time_operation('datetime', 'symmetric_difference')
-        38.8±3ms       30.0±0.6ms     0.77  arithmetic.IrregularOps.time_add
-       635±200μs          490±7μs     0.77  arithmetic.OffsetArrayArithmetic.time_add_dti_offset(<Day>)
-     7.32±0.05ms       5.59±0.2ms     0.76  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 8000, 0)
-     2.16±0.01ms      1.63±0.08ms     0.75  series_methods.NanOps.time_func('prod', 1000000, 'int8')
-     3.12±0.06ms       2.34±0.2ms     0.75  arithmetic.MixedFrameWithSeriesAxis.time_frame_op_with_series_axis0('sub')
-      7.10±0.1ms      5.13±0.07ms     0.72  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 2000, 0)
-      1.71±0.1ms      1.24±0.01ms     0.72  frame_methods.Iteration.time_items_cached
-     7.06±0.06ms      5.08±0.08ms     0.72  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 1000, 0)
-        23.2±3ms      16.6±0.07ms     0.72  frame_ctor.FromDictwithTimestamp.time_dict_with_timestamp_offsets(<Nano>)
-       51.5±30ms         36.6±1ms     0.71  gil.ParallelGroupbyMethods.time_loop(2, 'var')
-        214±20ms        152±0.8ms     0.71  frame_ctor.FromDicts.time_nested_dict_int64
-        132±80μs         92.4±6μs     0.70  ctors.SeriesConstructors.time_series_constructor(<function no_change at 0x7f8dd9d97940>, True, 'int')
-      1.42±0.3ms          991±9μs     0.70  dtypes.SelectDtypes.time_select_dtype_int_exclude('Int64')
-       718±200μs          500±8μs     0.70  categoricals.Constructor.time_interval
-     5.92±0.06ms      4.03±0.08ms     0.68  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 8000, 0)
-     6.40±0.09ms      4.26±0.05ms     0.67  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 1000, -2)
-     5.63±0.02ms      3.74±0.05ms     0.66  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 1000, 0)
-     1.73±0.02ms      1.14±0.08ms     0.66  series_methods.NanOps.time_func('sum', 1000000, 'int8')
-      6.17±0.4ms      4.02±0.06ms     0.65  arithmetic.IndexArithmetic.time_divide('int')
-      5.79±0.1ms      3.77±0.04ms     0.65  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 2000, 0)
-       95.3±40ms         61.6±2ms     0.65  gil.ParallelGroupbyMethods.time_loop(4, 'prod')
-       50.1±20ms       31.1±0.5ms     0.62  gil.ParallelGroupbyMethods.time_loop(2, 'min')
-      5.13±0.4ms      2.97±0.03ms     0.58  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 1000, -2)
-      4.02±0.5ms       2.30±0.1ms     0.57  arithmetic.IntFrameWithScalar.time_frame_op_with_scalar(<class 'numpy.int64'>, 5.0, <built-in function le>)
-       831±100μs         453±10μs     0.54  arithmetic.NumericInferOps.time_add(<class 'numpy.uint16'>)
-        763±10μs        414±100μs     0.54  indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Int8Engine'>, <class 'numpy.int8'>), 'monotonic_incr')
-        774±10μs         342±10μs     0.44  indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Int16Engine'>, <class 'numpy.int16'>), 'monotonic_incr')
-      26.1±0.5ms       8.32±0.3ms     0.32  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 2000, 2)
-      25.9±0.2ms       7.93±0.3ms     0.31  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 2000, -2)
-      24.0±0.1ms       6.87±0.2ms     0.29  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 2000, 2)
-      24.5±0.2ms       6.70±0.2ms     0.27  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 2000, -2)
-        92.1±1ms        130±0.9μs     0.00  indexing.NumericSeriesIndexing.time_loc_scalar(<class 'pandas.core.indexes.numeric.Float64Index'>, 'nonunique_monotonic_inc')
```


--------------------------------------------------------------------------------
**Other comparisons**

**Linear probing**
```
       before           after         ratio
     [c852b2ee]       [1dc7af75]
!      96.5±0.5ms           failed      n/a  hash_functions.IsinWithArange.time_isin(<class 'object'>, 8000, -2)
!         110±1ms           failed      n/a  hash_functions.IsinWithArange.time_isin(<class 'object'>, 8000, 2)
+         109±1ms       12.8±0.01s   117.14  hash_functions.IsinWithArange.time_isin(<class 'object'>, 2000, -2)
+      66.9±0.8ms       7.45±0.01s   111.28  hash_functions.IsinWithArange.time_isin(<class 'object'>, 1000, 2)
+         143±1ms       14.9±0.04s   104.27  hash_functions.IsinWithArange.time_isin(<class 'object'>, 2000, 2)
+        73.5±2ms       6.40±0.01s    87.12  hash_functions.IsinWithArange.time_isin(<class 'object'>, 1000, -2)
+      24.2±0.9ms          611±2ms    25.30  hash_functions.IsinWithArange.time_isin(<class 'numpy.int64'>, 1000, 2)
+        25.9±1ms          614±4ms    23.66  hash_functions.IsinWithArange.time_isin(<class 'numpy.uint64'>, 1000, 2)
+         520±6ms       11.8±0.03s    22.73  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 750000)
+      21.1±0.1ms         394±10ms    18.68  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 80000)
+         628±4ms       11.1±0.07s    17.60  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 900000)
+      16.0±0.4ms         281±10ms    17.59  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 70000)
+        480±10ms       8.15±0.05s    16.97  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 750000)
+     1.84±0.03ms       27.8±0.2ms    15.07  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 8000)
+         348±8μs      4.87±0.06ms    13.97  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 1300)
+     1.64±0.02ms       22.7±0.3ms    13.85  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 7000)
+         538±7μs       6.96±0.1ms    12.93  hash_functions.IsinWithRandomFloat.time_isin_outside(<class 'object'>, 2000)
+         683±4ms       8.59±0.06s    12.59  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 900000)
+      22.4±0.3ms          263±9ms    11.75  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 80000)
+     1.90±0.02ms       21.9±0.6ms    11.50  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 8000)
+      17.8±0.5ms          194±4ms    10.88  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 70000)
+     1.69±0.02ms       17.9±0.2ms    10.58  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 7000)
+         332±6μs      3.42±0.02ms    10.32  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 1300)
+         537±5μs      5.41±0.03ms    10.08  hash_functions.IsinWithRandomFloat.time_isin(<class 'object'>, 2000)
+      33.6±0.4ms          145±1ms     4.31  io.csv.ReadCSVConcatDatetime.time_read_csv
+     6.47±0.05ms       26.5±0.1ms     4.09  join_merge.Merge.time_merge_dataframe_integer_2key(False)
+      94.0±0.6ms          324±4ms     3.44  multiindex_object.GetLoc.time_large_get_loc
+      16.2±0.1ms       53.1±0.5ms     3.28  join_merge.Merge.time_merge_dataframe_integer_2key(True)
+       111±0.8ms          340±2ms     3.06  multiindex_object.GetLoc.time_large_get_loc_warm
+         662±8μs      1.17±0.04ms     1.77  groupby.GroupByMethods.time_dtype_as_field('datetime', 'nunique', 'transformation')
+     7.23±0.06ms       12.6±0.6ms     1.74  io.csv.ReadUint64Integers.time_read_uint64_neg_values
…